requirements.txt
pyTelegramBotAPI==4.23.0
elevenlabs>=1.7.0
python-dotenv>=1.0.1
openai>=2.2.0
httpx>=0.27.0
